// clients/accountingService/models/LoanCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LoanCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    loanTimestamp?: string;
    paymentDeadline?: string;
    value?: number;
    interestRate?: number;
    isCompundInterestRate?: boolean;
    loanTypeId?: string | null;
    currencyId?: string | null;
    tenantId?: string | null;
    enrolmentId?: string | null;
};